refactor!: change data type of probabilityPercentage prop of PrecipitationProbability class to Int

+++ app/src/main/java/com/example/justweather/domain/models/PrecipitationProbability.kt
@@ -11,7 +11,7 @@ data class PrecipitationProbability(
     val longitude:String,
     val hour: Int,
     val isAM:Boolean,
-    val probabilityPercentage: Float
+    val probabilityPercentage: Int
 )
 
 /**
